+++ src/core/Category/domain/Category.ts
@@ -1,9 +1,9 @@
 import { AggregateRoot } from '../../Shared/domain/AggregateRoot';
-import { TemplateId } from '../../Template/domain/TemplateId';
+import { TemplateId } from '../../Template/domain/value-object/TemplateId';
 import { CategoryId } from './value-object/CategoryId';
 import { CategoryName } from './value-object/CategoryName';
 import { CategoryShortDescription } from './value-object/CategoryShortDescription';
-import { Primitives } from '../../Shared/domain/common/Primitives';
+import { Primitives } from '../../Shared/domain/Primitives';
 
 export class Category extends AggregateRoot {
   constructor(
