+++ src/core/Contact/domain/Contact.ts
@@ -1,0 +1,11 @@
+import { AggregateRoot } from '../../Shared/domain/AggregateRoot';
+
+export class Contact extends AggregateRoot {
+  constructor() {
+    super();
+  }
+
+  toPrimitives() {
+    throw new Error('Method not implemented.');
+  }
+}
