+++ src/core/Message/domain/Message.ts
@@ -1,0 +1,11 @@
+import { AggregateRoot } from '../../Shared/domain/AggregateRoot';
+
+export class Message extends AggregateRoot {
+  constructor() {
+    super();
+  }
+
+  toPrimitives() {
+    throw new Error('Method not implemented.');
+  }
+}
