+++ src/core/Message/domain/value-object/MessageId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class MessageId extends Uuid {}
