+++ src/core/Shared/domain/Phone/Phone.ts
@@ -1,0 +1,10 @@
+import { AggregateRoot } from '../AggregateRoot';
+
+export class Phone extends AggregateRoot {
+  toPrimitives() {
+    throw new Error('Method not implemented.');
+  }
+  constructor() {
+    super();
+  }
+}
