+++ src/core/Shared/domain/Phone/PhoneId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../value-object/Uuid';
+
+export class PhoneId extends Uuid {}
