--- src/core/Shared/domain/common/PhoneNumber.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../value-object/StringValueObject';
-
-export class PhoneNumber extends StringValueObject {}
