--- src/core/Shared/domain/common/UserValueObject.ts
@@ -1,2 +1,0 @@
-import { StringValueObject } from '../value-object/StringValueObject';
-export class UserValueObject extends StringValueObject {}
