+++ src/core/Shared/domain/value-object/UserValueObject.ts
@@ -1,0 +1,2 @@
+import { StringValueObject } from './StringValueObject';
+export class UserValueObject extends StringValueObject {}
