--- src/core/Template/domain/TemplateId.ts
@@ -1,3 +1,0 @@
-import { Uuid } from '../../Shared/domain/value-object/Uuid';
-
-export class TemplateId extends Uuid {}
