--- src/core/Template/domain/TemplateName.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-export class TemplateName extends StringValueObject {}
