--- src/core/Template/domain/TemplatePreview.ts
@@ -1,2 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-export class TemplatePreview extends StringValueObject {}
