--- src/core/Template/domain/TemplateShortDescription.ts
@@ -1,2 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-export class TemplateShortDescription extends StringValueObject {}
