--- src/core/Template/domain/TemplateVariable.ts
@@ -1,4 +1,0 @@
-import { StringValueObject } from '../../Shared/domain/value-object/StringValueObject';
-
-// export class
-export class TemplateVariable extends StringValueObject {}
