+++ src/core/Template/domain/value-object/TemplateId.ts
@@ -1,0 +1,3 @@
+import { Uuid } from '../../../Shared/domain/value-object/Uuid';
+
+export class TemplateId extends Uuid {}
