+++ src/core/Template/domain/value-object/TemplateName.ts
@@ -1,0 +1,3 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+export class TemplateName extends StringValueObject {}
