+++ src/core/Template/domain/value-object/TemplatePreview.ts
@@ -1,0 +1,2 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+export class TemplatePreview extends StringValueObject {}
