+++ src/core/Template/domain/value-object/TemplateVariable.ts
@@ -1,0 +1,4 @@
+import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
+
+// export class
+export class TemplateVariable extends StringValueObject {}
